aggregator/telco-router/service/src/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
oauthlib==3.2.2
pymongo==4.3.3
python-dateutil==2.8.2
requests[security]==2.28.2
requests-mock==1.10.0
jwcrypto==1.4.2
pyOpenSSL==23.0.0
cryptography==39.0.2
ujson==5.7.0
jsonschema==4.17.3
rfc3987==1.3.8
phonenumbers==8.13.7
cachetools==5.3.0
retry==0.9.2
gunicorn==20.1.0
freezegun==1.2.2
coverage==7.2.2
dependency-check==0.6.0
